refactor(reports): move the drawing and the Word reports out of beam.py
Phase 3, second half. 782 more lines moved by the same cut-the-text script used
for the report tables: the seven section-drawing methods to mento/plots.py, and
the two Word-report methods to mento/documents.py. beam.plot() and
beam.shear_results_detailed_doc() stay as one-line delegations, so nothing
calling them changed.

  beam.py           2135 -> 1441
  report_tables.py  986
  plots.py          669
  documents.py      217

Neither new module is in mypy's ignore_errors list, which surfaced 36 real
typing errors in code that had never been checked: an undeclared _fig
attribute, three functions with no annotations, Optional axes and settings, and
detail dictionaries inferred as dict[str, object]. All fixed at the source -
_fig is declared beside _ax on RectangularSection, and the detail dictionaries
are typed where they are built - rather than by adding another ignore entry.

The two tests that called _format_rebar_layer_text as a beam method now import
it from mento.plots.

Still remaining in this phase: beam.py imports matplotlib and docx
transitively, because results.py supplies Formatter and TablePrinter for the
notebook views and imports both. Splitting results.py so a calculation-only
path exists is the last piece. The wall report tables have not moved either.

Suite 959 passed, coverage still 100%.
